Factor out a modules module
