+++ src/wasm-lib/kcl/src/execution/exec_ast.rs
@@ -10,16 +10,17 @@ use crate::{
         annotations,
         cad_op::{OpArg, Operation},
         state::ModuleState,
-        BodyType, ExecState, ExecutorContext, KclValue, MemoryFunction, Metadata, ModulePath, ModuleRepr,
-        ProgramMemory, TagEngineInfo, TagIdentifier,
+        BodyType, ExecState, ExecutorContext, KclValue, MemoryFunction, Metadata, ProgramMemory, TagEngineInfo,
+        TagIdentifier,
     },
+    modules::{ModuleId, ModulePath, ModuleRepr},
     parsing::ast::types::{
         ArrayExpression, ArrayRangeExpression, BinaryExpression, BinaryOperator, BinaryPart, BodyItem, CallExpression,
         CallExpressionKw, Expr, FunctionExpression, IfExpression, ImportPath, ImportSelector, ItemVisibility,
         LiteralIdentifier, LiteralValue, MemberExpression, MemberObject, Node, NodeRef, NonCodeValue, ObjectExpression,
         PipeExpression, TagDeclarator, UnaryExpression, UnaryOperator,
     },
-    source_range::{ModuleId, SourceRange},
+    source_range::SourceRange,
     std::{
         args::{Arg, KwArgs},
         FunctionKind,
