+++ src/wasm-lib/kcl/src/parsing/math.rs
@@ -131,7 +131,7 @@ mod tests {
             ast::types::{Literal, LiteralValue},
             token::NumericSuffix,
         },
-        source_range::ModuleId,
+        ModuleId,
     };
 
     #[test]
